Support block data with missing image_scales, after image upload
image_scales={}, missing default_field.

This happens right after an image is uploaded. Although, this is
probably a bug on the back-end, because at this time the scales are
already available, so it would be expected to also have the
image_scales in the payload.

Fall back to showing the original image, `@@image/image` in this case.